+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 2.0.0
+
+* **Breaking**: Make single file selection default
+* Add `multiple` option
+* Export the `ERR_ABORTED` error instance
+* Update docs
+
 ### 1.0.2
 
 * Switch from babel to typescript